Reorganize argument parsing
Arg parsing is now split into common options (in bin/documentation.js)
and command-specific options (in each command module, exported as
parseArgs).

This includes moving `--private` and `--github` to the common args
instead of having them under the build/serve commands.  Reason: these
two args are options understood by the main module, as opposed to args
like `--format` which are only understood by the actual output piece
in build/serve.

Also removes `args.js`, which had previously exported the argument
parsing for use in situations like documentation-readme.  The need
for this, I think, is largely made irrelevant in the face of slightly
generalizing the CLI to use commands.
