+++ lib/commands/lint.js
@@ -4,6 +4,10 @@
 
 module.exports = lint;
 module.exports.description = 'check for common style and uniformity mistakes'
+module.exports.parseArgs = function (yargs) {
+  yargs.example('$0 lint project.js', 'check documentation style')
+  .help('help')
+}
 
 function lint(documentation, parsedArgs) {
   documentation.lint(parsedArgs.inputs, parsedArgs.options, function (err, lintOutput) {
