+++ lib/commands/serve.js
@@ -7,6 +7,7 @@ var build = require('./build')
 
 module.exports = serve
 module.exports.description = 'generate, update, and display HTML documentation'
+module.exports.parseArgs = build.addOutputArgs
 
 var server = new Server();
 server.on('listening', function () {
@@ -14,6 +15,7 @@ server.on('listening', function () {
 });
 
 function serve(documentation, parsedArgs) {
+  parsedArgs.commandOptions.format = 'html'
   build(documentation, parsedArgs, function (err, output) {
     if (err) {
       return server.setFiles([errorPage(err)]).start();
